requirements.txt
celery==3.1.18
docker-py==1.3.1
Flask==0.10.1
Flask-Script==2.0.5
Flask-SQLAlchemy==2.0
Flask-Security==1.7.4
redis==2.10.3

